refactor(cli): split into modules
