+++ openspec/changes/refactor-cli-modules/README.md
@@ -1,0 +1,3 @@
+# refactor-cli-modules
+
+Split the large CLI implementation into focused modules without changing behavior.
