+++ openspec/changes/refactor-cli-modules/specs/agentpack-cli/spec.md
@@ -1,0 +1,10 @@
+# agentpack-cli (delta)
+
+## ADDED Requirements
+
+### Requirement: CLI implementation is modular
+The system SHALL structure the CLI implementation as a set of focused modules (e.g., `src/cli/commands/*`) rather than a single monolithic file, so command behavior and output contracts remain easier to maintain and review.
+
+#### Scenario: Command handlers are localized
+- **WHEN** a developer adds or updates a CLI subcommand handler
+- **THEN** the change is confined to a dedicated module under `src/cli/commands/` and shared helpers under `src/cli/`
