+++ src/cli/mod.rs
@@ -1,0 +1,8 @@
+mod args;
+mod commands;
+mod dispatch;
+mod human;
+mod json;
+mod util;
+
+pub use dispatch::run;
